+++ apps/sandbox/app/[root]/_layout.tsx
@@ -7,7 +7,7 @@ export const unstable_settings = {
 
 export default function RootLayout() {
   return (
-    <Stack screenOptions={{ headerShown: false }}>
+    <Stack>
       <Stack.Screen name="index" />
     </Stack>
   );
